Moved responses from model to responses
