+++ src/digitalIdentity-backend/src/main/java/didentity/amos/digitalIdentity/services/DIConnectionService.java
@@ -11,8 +11,8 @@ import org.springframework.web.client.RestClientException;
 
 import didentity.amos.digitalIdentity.enums.UserRole;
 import didentity.amos.digitalIdentity.model.Connection;
-import didentity.amos.digitalIdentity.model.ConnectionsResponse;
 import didentity.amos.digitalIdentity.model.Content;
+import didentity.amos.digitalIdentity.messages.responses.ConnectionsResponse;
 import didentity.amos.digitalIdentity.messages.responses.CreateConnectionResponse;
 import didentity.amos.digitalIdentity.model.User;
 import didentity.amos.digitalIdentity.repository.UserRepository;
